Move mandatory cols to new file
