+++ tests/test_file_parser/test_parse_internal_continuous.py
@@ -6,7 +6,7 @@ from numpy import ndarray, dtype
 
 from gaiaxpy.core.satellite import BANDS
 from gaiaxpy.file_parser.parse_internal_continuous import InternalContinuousParser
-from gaiaxpy.input_reader.file_reader import MANDATORY_COLS
+from gaiaxpy.input_reader.mandatory_columns import MANDATORY_COLS
 from tests.files.paths import mean_spectrum_avro_file, mean_spectrum_csv_file, mean_spectrum_fits_file,\
     mean_spectrum_xml_file
 from tests.utils.utils import get_spectrum_with_source_id
