+++ tests/test_generator/test_synthetic_photometry_generator.py
@@ -5,7 +5,7 @@ from gaiaxpy.core.satellite import BANDS
 from gaiaxpy.file_parser.parse_internal_continuous import InternalContinuousParser
 from gaiaxpy.generator.photometric_system import PhotometricSystem
 from gaiaxpy.generator.synthetic_photometry_generator import _generate_synthetic_photometry
-from gaiaxpy.input_reader.file_reader import MANDATORY_COLS
+from gaiaxpy.input_reader.mandatory_columns import MANDATORY_COLS
 from gaiaxpy.spectrum.sampled_basis_functions import SampledBasisFunctions
 from gaiaxpy.spectrum.single_synthetic_photometry import SingleSyntheticPhotometry
 from tests.files.paths import mean_spectrum_avro_file, mean_spectrum_csv_file, mean_spectrum_xml_file, \
