+++ tests/test_input_reader/test_input_reader_missing_bp.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pandas.testing as pdt
 
 from gaiaxpy import convert
-from gaiaxpy.input_reader.file_reader import MANDATORY_COLS
+from gaiaxpy.input_reader.mandatory_columns import MANDATORY_COLS
 from gaiaxpy.input_reader.input_reader import InputReader
 from tests.files.paths import with_missing_bp_csv_file, with_missing_bp_ecsv_file, with_missing_bp_fits_file,\
     with_missing_bp_xml_file, with_missing_bp_xml_plain_file, input_reader_solution_path
